front_end_integration/react_flask/api/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pyjwt==2.8.0
python-dotenv==1.0.0
werkzeug==2.3.7
